Consolidate animation controls into a single module.
