+++ src/osc/animation.rs
@@ -2,13 +2,17 @@ use rosc::OscMessage;
 use tunnels::clock_bank::{ClockIdxExt, N_CLOCKS};
 
 use crate::animation::ControlMessage::Animation as WrapAnimation;
-use crate::fixture::FixtureControlMessage;
+use crate::fixture::{FixtureControlMessage, N_ANIM};
 use crate::osc::HandleStateChange;
 use crate::osc::{send_button, send_float, ControlMap, MapControls, RadioButton};
 
 use tunnels::animation::{ControlMessage, StateChange, Waveform::*};
 
+use super::label_array::LabelArray;
+
 const GROUP: &str = "Animation";
+
+// Base animation system
 
 // knobs
 const SPEED: &str = "Speed";
@@ -43,6 +47,7 @@ const CLOCK_SOURCE: RadioButton = RadioButton {
     n: N_CLOCKS + 1,
     x_primary_coordinate: false,
 };
+
 pub struct AnimationControls;
 
 impl MapControls for AnimationControls {
@@ -102,6 +107,67 @@ impl MapControls for AnimationControls {
             USE_AUDIO_SIZE,
             FixtureAnimation(WrapAnimation(ToggleUseAudioSize)),
         );
+
+        TargetAndSelectControls.map_controls(map);
+    }
+}
+
+// Targeting/selection
+
+const N_ANIM_TARGET: usize = 11;
+
+const ANIMATION_SELECT: RadioButton = RadioButton {
+    group: GROUP,
+    control: "Select",
+    n: N_ANIM,
+    x_primary_coordinate: false,
+};
+
+const ANIMATION_TARGET_SELECT: RadioButton = RadioButton {
+    group: GROUP,
+    control: "Target",
+    n: N_ANIM_TARGET,
+    x_primary_coordinate: false,
+};
+
+const ANIMATION_LABELS: LabelArray = LabelArray {
+    group: GROUP,
+    control: "TargetLabel",
+    n: N_ANIM_TARGET,
+    empty_label: "XXXXXX",
+};
+
+struct TargetAndSelectControls;
+
+impl MapControls for TargetAndSelectControls {
+    fn map_controls(&self, map: &mut ControlMap<FixtureControlMessage>) {
+        use crate::animation::ControlMessage;
+        use FixtureControlMessage::Animation;
+
+        map.add_radio_button_array(ANIMATION_TARGET_SELECT, |msg| {
+            Animation(ControlMessage::Target(msg))
+        });
+        map.add_radio_button_array(ANIMATION_SELECT, |msg| {
+            Animation(ControlMessage::Select(msg))
+        });
+    }
+}
+
+impl HandleStateChange<crate::animation::StateChange> for AnimationControls {
+    fn emit_state_change<S>(sc: crate::animation::StateChange, send: &mut S)
+    where
+        S: FnMut(OscMessage),
+    {
+        match sc {
+            crate::animation::StateChange::Animation(msg) => {
+                AnimationControls::emit_state_change(msg, send)
+            }
+            crate::animation::StateChange::Select(msg) => ANIMATION_SELECT.set(msg, send),
+            crate::animation::StateChange::Target(msg) => ANIMATION_TARGET_SELECT.set(msg, send),
+            crate::animation::StateChange::Labels(labels) => {
+                ANIMATION_LABELS.set(labels.into_iter(), send)
+            }
+        }
     }
 }
